[Enhancement] Support manual deletion of PersistentIndexMeta (#9888)
To avoid the bug of ignoring the persistent index, we do not support rebuilding the persistent index for the time being When an error is found in the persistent index.
However, this makes it a bit difficult to recover the persistent index. To simplify the recovery process, add support for deleting persistent index meta.